backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
marshmallow==3.20.1
scikit-learn==1.6.1
numpy==1.26.4
pandas==2.2.3
PyPDF2==3.0.1
joblib==1.4.2
scipy==1.11.4
requests==2.31.0
reportlab==4.0.7
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
redis==5.0.1
